# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Finca Farm Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /farms/{farmId}/inventory:
    get:
      operationId: listInventoryItems
      tags: [inventory]
      summary: List inventory items
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Add inventory item
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemRequest"
      responses:
        "201":
          description: Created inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /farms/{farmId}/inventory/{itemId}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get inventory item details
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItemDetail"
    put:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryItemRequest"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /farms/{farmId}/inventory/{itemId}/log:
    post:
      operationId: logInventoryAction
      tags: [inventory]
      summary: Log inventory action (use/restock/adjust)
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogInventoryActionRequest"
      responses:
        "201":
          description: Log entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryLog"
...
  /farms/{farmId}/zones:
    get:
      operationId: listZones
      tags: [zones]
      summary: List zones for a farm
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Zone"
    post:
      operationId: createZone
      tags: [zones]
      summary: Create a zone
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateZoneRequest"
      responses:
        "201":
          description: Created zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
...
components:
  schemas:
...
    InventoryItem:
      type: object
      properties:
        id:
          type: string
        farmId:
          type: string
        category:
          type: string
          enum: [feed, medicine, tools, supplies]
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
          enum: [kg, liters, units, bags, doses, other]
        lowStockThreshold:
          type: number
        costPerUnitCop:
          type: number
        supplierName:
          type: string
        supplierContact:
          type: string
        expirationDate:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [ok, low, expired]
        createdAt:
          type: string
      required:
        - id
        - farmId
        - category
        - name
        - quantity
        - unit
    InventoryItemDetail:
      allOf:
        - $ref: "#/components/schemas/InventoryItem"
        - type: object
          properties:
            logs:
              type: array
              items:
                $ref: "#/components/schemas/InventoryLog"
    CreateInventoryItemRequest:
      type: object
      properties:
        category:
          type: string
          enum: [feed, medicine, tools, supplies]
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
          enum: [kg, liters, units, bags, doses, other]
        lowStockThreshold:
          type: number
        costPerUnitCop:
          type: number
        supplierName:
          type: string
        supplierContact:
          type: string
        expirationDate:
          type: string
        notes:
          type: string
      required:
        - category
        - name
        - quantity
        - unit
    UpdateInventoryItemRequest:
      type: object
      properties:
        category:
          type: string
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        lowStockThreshold:
          type: number
        costPerUnitCop:
          type: number
        supplierName:
          type: string
        supplierContact:
          type: string
        expirationDate:
          type: string
        notes:
          type: string
    LogInventoryActionRequest:
      type: object
      properties:
        action:
          type: string
          enum: [added, used, adjusted, expired]
        quantityChange:
          type: number
        animalId:
          type: string
        notes:
          type: string
      required:
        - action
        - quantityChange
    InventoryLog:
      type: object
      properties:
        id:
          type: string
        itemId:
          type: string
        action:
          type: string
        quantityChange:
          type: number
        animalId:
          type: string
        notes:
          type: string
        createdAt:
          type: string
      required:
        - id
        - itemId
        - action
        - quantityChange
...
    Zone:
      type: object
      properties:
        id:
          type: string
        farmId:
          type: string
        name:
          type: string
        zoneType:
          type: string
          enum: [grazing, feeding, water, waste, crops, shelter, storage, other]
        color:
          type: string
        capacity:
          type: integer
        areaHectares:
          type: number
        notes:
          type: string
        createdAt:
          type: string
      required:
        - id
        - farmId
        - name
    CreateZoneRequest:
      type: object
      properties:
        name:
          type: string
        zoneType:
          type: string
        color:
          type: string
        capacity:
          type: integer
        areaHectares:
          type: number
        notes:
          type: string
      required:
        - name
...