# lib/api-zod/src/generated/types/createInventoryItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finca Farm Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateInventoryItemRequestCategory } from "./createInventoryItemRequestCategory";
import type { CreateInventoryItemRequestUnit } from "./createInventoryItemRequestUnit";

export interface CreateInventoryItemRequest {
  category: CreateInventoryItemRequestCategory;
  name: string;
  quantity: number;
  unit: CreateInventoryItemRequestUnit;
  lowStockThreshold?: number;
  costPerUnitCop?: number;
  supplierName?: string;
  supplierContact?: string;
  expirationDate?: string;
  notes?: string;
}