# lib/api-zod/src/generated/types/logInventoryActionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finca Farm Management API
 * OpenAPI spec version: 0.1.0
 */
import type { LogInventoryActionRequestAction } from "./logInventoryActionRequestAction";

export interface LogInventoryActionRequest {
  action: LogInventoryActionRequestAction;
  quantityChange: number;
  animalId?: string;
  notes?: string;
}