# lib/api-zod/src/generated/types/updateInventoryItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finca Farm Management API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateInventoryItemRequest {
  category?: string;
  name?: string;
  quantity?: number;
  unit?: string;
  lowStockThreshold?: number;
  costPerUnitCop?: number;
  supplierName?: string;
  supplierContact?: string;
  expirationDate?: string;
  notes?: string;
}